+++ firehose/info/info_filler.go
@@ -8,15 +8,33 @@ import (
 	pbfirehose "github.com/streamingfast/pbgo/sf/firehose/v2"
 )
 
-var DefaultInfoResponseFiller = func(firstStreamableBlock *pbbstream.Block, resp *pbfirehose.InfoResponse) error {
+var DefaultInfoResponseFiller = func(firstStreamableBlock *pbbstream.Block, resp *pbfirehose.InfoResponse, validate bool) error {
 	resp.FirstStreamableBlockId = firstStreamableBlock.Id
+
+	for _, protocol := range wellknown.WellKnownProtocols {
+		if protocol.BlockType == firstStreamableBlock.Payload.TypeUrl {
+			resp.BlockIdEncoding = protocol.BytesEncoding
+			break
+		}
+	}
+
+	if !validate {
+		if resp.ChainName == "" {
+			// still try to fill the chain name if it is not given
+			if chain := wellknown.WellKnownProtocols.ChainByGenesisBlock(firstStreamableBlock.Number, firstStreamableBlock.Id); chain != nil {
+				resp.ChainName = chain.Name
+				resp.ChainNameAliases = chain.Aliases
+			}
+		}
+		return nil
+	}
 
 	if resp.ChainName != "" {
 		if chain := wellknown.WellKnownProtocols.ChainByName(resp.ChainName); chain != nil {
 			if firstStreamableBlock.Number == chain.GenesisBlockNumber && chain.GenesisBlockID != firstStreamableBlock.Id { // we don't check if the firstStreamableBlock is something other than our well-known genesis block
 				return fmt.Errorf("chain name defined in flag: %q inconsistent with the genesis block ID %q (expected: %q)", resp.ChainName, ox(firstStreamableBlock.Id), ox(chain.GenesisBlockID))
 			}
-			resp.ChainName = chain.Name
+			resp.ChainName = chain.Name // ensure we use the canonical name if the user provided one of the aliases
 			resp.ChainNameAliases = chain.Aliases
 		} else if chain := wellknown.WellKnownProtocols.ChainByGenesisBlock(firstStreamableBlock.Number, firstStreamableBlock.Id); chain != nil {
 			return fmt.Errorf("chain name defined in flag: %q inconsistent with the one discovered from genesis block %q", resp.ChainName, chain.Name)
@@ -25,13 +43,6 @@ var DefaultInfoResponseFiller = func(firstStreamableBlock *pbbstream.Block, resp
 		if chain := wellknown.WellKnownProtocols.ChainByGenesisBlock(firstStreamableBlock.Number, firstStreamableBlock.Id); chain != nil {
 			resp.ChainName = chain.Name
 			resp.ChainNameAliases = chain.Aliases
-		}
-	}
-
-	for _, protocol := range wellknown.WellKnownProtocols {
-		if protocol.BlockType == firstStreamableBlock.Payload.TypeUrl {
-			resp.BlockIdEncoding = protocol.BytesEncoding
-			break
 		}
 	}